tyrquake: widen CL_KeepaliveMessage's host_time storage to double
CL_KeepaliveMessage maintains a static fp32 'lastmsg' marker that
snapshots host_time, plus a fp32 local 'time' for the current
host_time value, then throttles to one keepalive per 5 simulated
seconds via 'time - lastmsg < 5'. host_time is a double that grows
monotonically since process start, so each fp32 snapshot drifts
the usual way -- ulp ~ 0.06 at host_time ~ 1e6 (~12 d). The 5 s
comparison's precision degrades correspondingly.

The fix here is the simplest one in the audit: both lastmsg and the
local 'time' are private to this function (lastmsg is a static
inside the function body, not a struct field), so just widen them
to double. No ABI surface, no callers affected, no countdown
semantics needed since the existing absolute-timestamp comparison
becomes precise once the storage is double-precision throughout.

This function is rarely exercised in the libretro context (it
early-returns when sv.active, which the local listenserver almost
always is), but exposed when connected to a remote network server.
Same pattern as the static float lastmsg that swept-but-missed
the prior audit cycles -- the static-float sweep walked every
file's statics and got every other one, but missed verifying the
ones inside function bodies. lastmsg was declared inside
CL_KeepaliveMessage rather than at file scope, so the
'static float \w+ ;' file-level grep didn't find it.

+++ common/cl_parse.c
@@ -226,8 +226,15 @@ so the server doesn't disconnect.
 static void
 CL_KeepaliveMessage(void)
 {
-    float time;
-    static float lastmsg;
+    /* host_time is double; widen both the static throttle marker
+     * and the local snapshot so the 5-second cooldown comparison
+     * stays precise after long engine sessions. lastmsg persists
+     * across map / server transitions, so its fp32 storage would
+     * accumulate the usual truncation error otherwise (~0.06 s
+     * at host_time ~ 1e6, ~12 d). No ABI surface -- both
+     * are local to this translation unit and this function. */
+    double time;
+    static double lastmsg;
     int ret;
     sizebuf_t old;
     byte olddata[8192];
